Fixed Bugs - SQLite Connection String service typo and wrong path to SQLite queries.
Didnt notice that Copilot suggested one useless private field and used it to assign connection string data in ctor instead of correct connection string service passed as parameter in ctor. And again I trusted Copilot that suggested method to create path to SQLite queries folder is correct but that path lead to currently opened app base directory instead of library app...

And I told everyone: Don't trust Copilot, don't trust Copilot, or you'll go crazy. But it's in vain, it's in vain, it's in vain!