Add API project
Add API project with necessary dependencies and services. Add Basic Authentication controller for Desktop project. API project is needed mainly to secure sensitive data in Desktop project just in case if we decide to share this project with anonymous users and not keep it as some Check-in terminals at entrance to company.